Remove dependency to TestLinkFixture from TestLinkAPI
TestLinkAPI has a dependency to TestLinkFixtureAttribute. The intention of
this atribute is to add it to existing testsuites in order to bind it to
concrete testlink project (i.e. connection (URL) + project data like test
plan, build, platform etc.)

This commit removes this depencencies by introducing two distinct classes
which identifies the connection (TestLinkConnectionData) and the testlink
project related data (TestLinkProjectData). This allows using the
TestLinkAPI project without a reference to the TestLinkFixture project.
This TestLinkFixture project is now only used in NUnit context, i.e.
testsuites can include this to add the TestLinkFixture attribute to.

In order to keep existing functionality there is now a conversion in
ResultExporter for the nunit addon which converts data from
TestLinkFixture attribute to TestLinkConnectionData and
TestLinkProjectData, respectively.