Implement caching inside TestLinkProjectData
caching of project data (testplan, project, build and platform) was so far
done in TestLinkAdaptor while calling UpdateConnectionData.
This logic is now moved to TestLinkProjectData using Directory class for
each project element. If there is no hit in the directory, the server is
contacted to get the information from the server.